# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Lourde Fitness Studio
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /classes:
    get:
      tags: [Classes]
      operationId: listClasses
      summary: List all classes in the weekly timetable
      responses:
        "200":
          description: List of classes.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClassSession"
    post:
      tags: [Classes]
      operationId: createClass
      summary: Create a class session (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClassInput"
      responses:
        "201":
          description: Created class.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClassSession"
        "403":
          description: Admins only.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /classes/{id}:
    patch:
      tags: [Classes]
      operationId: updateClass
      summary: Update a class session (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClassUpdate"
      responses:
        "200":
          description: Updated class.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClassSession"
        "403":
          description: Admins only.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Class not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    delete:
      tags: [Classes]
      operationId: deleteClass
      summary: Delete a class session (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted.
        "403":
          description: Admins only.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Class not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  # ---------- Packages / Promos ----------
  /packages:
    get:
      tags: [Packages]
      operationId: listPackages
      summary: List all packages/promos
      parameters:
        - name: activeOnly
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: List of packages.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Package"
    post:
      tags: [Packages]
      operationId: createPackage
      summary: Create a package/promo (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PackageInput"
      responses:
        "201":
          description: Created package.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Package"
        "403":
          description: Admins only.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /packages/{id}:
    patch:
      tags: [Packages]
      operationId: updatePackage
      summary: Update a package/promo (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PackageUpdate"
      responses:
        "200":
          description: Updated package.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Package"
        "403":
          description: Admins only.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Package not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    delete:
      tags: [Packages]
      operationId: deletePackage
      summary: Delete a package/promo (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted.
        "403":
          description: Admins only.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Package not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  # ---------- Payment QR codes ----------
  /payment-methods:
    get:
      tags: [PaymentMethods]
      operationId: listPaymentMethods
      summary: List active payment QR codes shown to students at checkout
      responses:
        "200":
          description: List of payment methods.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PaymentMethod"
    post:
      tags: [PaymentMethods]
      operationId: createPaymentMethod
      summary: Add a payment QR code (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentMethodInput"
      responses:
        "201":
          description: Created payment method.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentMethod"
        "403":
          description: Admins only.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /payment-methods/{id}:
    patch:
      tags: [PaymentMethods]
      operationId: updatePaymentMethod
      summary: Update a payment QR code (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentMethodUpdate"
      responses:
        "200":
          description: Updated payment method.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentMethod"
        "403":
          description: Admins only.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Payment method not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    delete:
      tags: [PaymentMethods]
      operationId: deletePaymentMethod
      summary: Delete a payment QR code (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted.
        "403":
          description: Admins only.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Payment method not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  # ---------- Purchases ----------
  /purchases/me:
    get:
      tags: [Purchases]
      operationId: listMyPurchases
      summary: List the current student's own purchases
      responses:
        "200":
          description: List of purchases.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Purchase"
        "401":
          description: Not signed in.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    post:
      tags: [Purchases]
      operationId: createPurchase
      summary: Submit a purchase with proof-of-payment screenshot (student)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchaseInput"
      responses:
        "201":
          description: Created purchase, pending review.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Purchase"
        "401":
          description: Not signed in.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /purchases:
    get:
      tags: [Purchases]
      operationId: listPurchases
      summary: List all purchases (admin only)
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/PurchaseStatus"
      responses:
        "200":
          description: List of purchases.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Purchase"
        "403":
          description: Admins only.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /purchases/{id}:
    patch:
      tags: [Purchases]
      operationId: reviewPurchase
      summary: Confirm or reject a purchase (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchaseReview"
      responses:
        "200":
          description: Updated purchase.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Purchase"
        "403":
          description: Admins only.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Purchase not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  # ---------- Studio rental reservations ----------
...
components:
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ---------- Storage ----------
...
    ClassSession:
      type: object
      required:
        [
          id,
          title,
          instructor,
          dayOfWeek,
          startTime,
          endTime,
          level,
          capacity,
          description,
          createdAt,
        ]
      properties:
        id:
          type: integer
        title:
          type: string
        instructor:
          type: string
        dayOfWeek:
          type: integer
          minimum: 0
          maximum: 6
          description: 0 = Sunday .. 6 = Saturday
        startTime:
          type: string
          description: "24h time, e.g. 18:00"
        endTime:
          type: string
          description: "24h time, e.g. 19:00"
        level:
          type: string
        capacity:
          type: integer
        description:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    ClassInput:
      type: object
      required: [title, instructor, dayOfWeek, startTime, endTime, level, capacity]
      properties:
        title:
          type: string
          minLength: 1
        instructor:
          type: string
          minLength: 1
        dayOfWeek:
          type: integer
          minimum: 0
          maximum: 6
        startTime:
          type: string
          minLength: 1
        endTime:
          type: string
          minLength: 1
        level:
          type: string
          minLength: 1
        capacity:
          type: integer
          minimum: 1
        description:
          type: string
    ClassUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        instructor:
          type: string
          minLength: 1
        dayOfWeek:
          type: integer
          minimum: 0
          maximum: 6
        startTime:
          type: string
          minLength: 1
        endTime:
          type: string
          minLength: 1
        level:
          type: string
          minLength: 1
        capacity:
          type: integer
          minimum: 1
        description:
          type: ["string", "null"]

    # ---------- Packages ----------
    Package:
      type: object
      required:
        [
          id,
          name,
          description,
          category,
          price,
          originalPrice,
          sessionCount,
          validityDays,
          imageObjectPath,
          isPromo,
          active,
          createdAt,
        ]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        category:
          type: string
          description: "e.g. Pole, Aerial, Dance, Open Studio"
        price:
          type: number
        originalPrice:
          type: ["number", "null"]
          description: Set when isPromo is true to show a struck-through price.
        sessionCount:
          type: integer
        validityDays:
          type: integer
        imageObjectPath:
          type: ["string", "null"]
        isPromo:
          type: boolean
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
    PackageInput:
      type: object
      required: [name, description, category, price, sessionCount, validityDays]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
          minLength: 1
        category:
          type: string
          minLength: 1
        price:
          type: number
          minimum: 0
        originalPrice:
          type: number
          minimum: 0
        sessionCount:
          type: integer
          minimum: 1
        validityDays:
          type: integer
          minimum: 1
        imageObjectPath:
          type: string
        isPromo:
          type: boolean
        active:
          type: boolean
    PackageUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
          minLength: 1
        category:
          type: string
          minLength: 1
        price:
          type: number
          minimum: 0
        originalPrice:
          type: ["number", "null"]
        sessionCount:
          type: integer
          minimum: 1
        validityDays:
          type: integer
          minimum: 1
        imageObjectPath:
          type: ["string", "null"]
        isPromo:
          type: boolean
        active:
          type: boolean

    # ---------- Payment methods ----------
    PaymentMethod:
      type: object
      required: [id, label, qrImageObjectPath, instructions, active, createdAt]
      properties:
        id:
          type: integer
        label:
          type: string
          description: "e.g. GCash, Bank Transfer"
        qrImageObjectPath:
          type: string
        instructions:
          type: ["string", "null"]
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
    PaymentMethodInput:
      type: object
      required: [label, qrImageObjectPath]
      properties:
        label:
          type: string
          minLength: 1
        qrImageObjectPath:
          type: string
          minLength: 1
        instructions:
          type: string
        active:
          type: boolean
    PaymentMethodUpdate:
      type: object
      properties:
        label:
          type: string
          minLength: 1
        qrImageObjectPath:
          type: string
          minLength: 1
        instructions:
          type: ["string", "null"]
        active:
          type: boolean

    # ---------- Purchases ----------
    PurchaseStatus:
      type: string
      enum: [pending, confirmed, rejected]
    Purchase:
      type: object
      required:
        [
          id,
          studentId,
          packageId,
          packageName,
          amount,
          screenshotObjectPath,
          status,
          reviewNotes,
          submittedAt,
          reviewedAt,
        ]
      properties:
        id:
          type: integer
        studentId:
          type: integer
        packageId:
          type: integer
        packageName:
          type: string
        amount:
          type: number
        screenshotObjectPath:
          type: string
        status:
          $ref: "#/components/schemas/PurchaseStatus"
        reviewNotes:
          type: ["string", "null"]
        submittedAt:
          type: string
          format: date-time
        reviewedAt:
          type: ["string", "null"]
          format: date-time
    PurchaseInput:
      type: object
      required: [packageId, amount, screenshotObjectPath]
      properties:
        packageId:
          type: integer
        amount:
          type: number
          minimum: 0
        screenshotObjectPath:
          type: string
          minLength: 1
    PurchaseReview:
      type: object
      required: [status]
      properties:
        status:
          $ref: "#/components/schemas/PurchaseStatus"
        reviewNotes:
          type: string

    # ---------- Rentals ----------
...