# lib/api-zod/src/generated/types/errorEnvelope.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Lourde Fitness Studio
 * OpenAPI spec version: 0.1.0
 */

export interface ErrorEnvelope {
  error: string;
}