# lib/api-zod/src/generated/types/packageUpdate.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Lourde Fitness Studio
 * OpenAPI spec version: 0.1.0
 */

export interface PackageUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  description?: string;
  /** @minLength 1 */
  category?: string;
  /** @minimum 0 */
  price?: number;
  /** @nullable */
  originalPrice?: number | null;
  /** @minimum 1 */
  sessionCount?: number;
  /** @minimum 1 */
  validityDays?: number;
  /** @nullable */
  imageObjectPath?: string | null;
  isPromo?: boolean;
  active?: boolean;
}