# lib/api-zod/src/generated/types/paymentMethod.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Lourde Fitness Studio
 * OpenAPI spec version: 0.1.0
 */

export interface PaymentMethod {
  id: number;
  /** e.g. GCash, Bank Transfer */
  label: string;
  qrImageObjectPath: string;
  /** @nullable */
  instructions: string | null;
  active: boolean;
  createdAt: Date;
}